+++ core/src/Authorization.html
@@ -68,7 +68,7 @@
   import * as AuthHelpers from './utilities/helpers/auth-helpers';
   import { oAuth2ImplicitGrant } from './providers/auth/oAuth2ImplicitGrant.js';
   import { openIdConnect } from './providers/auth/openIdConnect.js';
-  import { TOP_NAV_DEFAULTS } from './utilities/constants';
+  import { TOP_NAV_DEFAULTS } from './utilities/luigi-config-defaults';
   import * as NavigationHelpers from './utilities/helpers/navigation-helpers';
 
   let idpProviderInstance;
