+++ core/src/services/routing.js
@@ -5,7 +5,7 @@ import * as RoutingHelpers from '../utilities/helpers/routing-helpers';
 import { LuigiConfig } from './config';
 import * as GenericHelpers from '../utilities/helpers/generic-helpers';
 import * as Iframe from './iframe';
-import { NAVIGATION_DEFAULTS } from '../utilities/constants';
+import { NAVIGATION_DEFAULTS } from './../utilities/luigi-config-defaults';
 
 export const getNodePath = (node, params) => {
   return node
